feat(main): add either a version command and version flag
